# Workr-backend/workr-swagger-api.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "Workr API",
    "version" : "2.1.0-oas3",
    "title" : "Workr API",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/user/password" : {
      "put" : {
        "tags" : [ "General" ],
        "summary" : "Change password",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "operationId" : "changePassword",
        "description" : "The user can change his password",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "password" ],
                "properties" : {
                  "newpassword" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "description" : "Returns the edited profile"
        }
      }
    },
    "/user/email" : {
      "put" : {
        "tags" : [ "General" ],
        "summary" : "Updated email",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "operationId" : "changeEmail",
        "description" : "changes the email of the user",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "email" ],
                "properties" : {
                  "email" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "description" : "Returns the edited profile"
        }
      }
    },
    "/profile" : {
      "get" : {
        "tags" : [ "General" ],
        "summary" : "Shows the user his profile",
        "operationId" : "profile",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "description" : "Shows the user his profile",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        }
      }
    },
    "/profile/{profileid}" : {
      "get" : {
        "tags" : [ "General" ],
        "summary" : "Shows the user the other profile",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "operationId" : "profileUser",
        "description" : "Shows the user the other profile",
        "parameters" : [ {
          "name" : "profileid",
          "in" : "path",
          "description" : "the profileid that you are requesting",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        }
      }
    },
...
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/workr/workrAPI/1.0.0"
  } ],
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    },
    "schemas" : {
      "User" : {
        "type" : "object",
        "required" : [ "userId", "username", "name", "email", "picture", "phone" ],
        "properties" : {
          "userId" : {
            "type" : "integer",
            "example" : 12
          },
          "name" : {
            "type" : "string",
            "example" : "John Doe"
          },
          "username" : {
            "type" : "string",
            "example" : "JohnDoe"
          },
          "email" : {
            "type" : "string",
            "example" : "johndoe@gmail.com"
          },
          "picture" : {
            "type" : "string",
            "example" : "base64 string"
          },
          "phone" : {
            "type" : "string",
            "example" : 6.12345678E8
          }
        }
      },
...
    },
    "responses" : {
      "usernotfound" : {
        "description" : "User not found"
      },
      "forbidden" : {
        "description" : "You are not logged in."
      }
    }
  }
}